pyproject.toml
[tool.poetry]
name = "file-processor-drf"
version = "0.1.0"
description = "Django REST API that allows you to upload files to the server and then process them asynchronously using Celery."
authors = ["ShadeSWD <seawolfy60@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "^69.0.3"
Django = "^5.0.2"
python-dotenv = "^1.0.1"
pre-commit = "^3.6.1"
flake8 = "^7.0.0"
black = "^24.1.1"
isort = "^5.13.2"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
celery = "^5.3.6"
eventlet = "^0.35.1"
redis = "^5.0.1"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
coverage = "^7.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
